Extract Encryption extension and make Link immutable
